somenews/requirements.txt
Django-5.1
django-filter-24.3
django-ckeditor-6.7.1
django-js-asset-2.2.0
python-dotenv-1.0.1
amqp-5.2.0
billiard-4.2.0
celery-5.4.0
click-8.1.7
click-didyoumean-0.3.1
click-plugins-1.1.1
click-repl-0.3.0
kombu-5.4.0
prompt-toolkit-3.0.47
python-dateutil-2.9.0.post0
six-1.16.0
tzdata-2024.1
vine-5.1.0
wcwidth-0.2.13
redis-5.0.8
cron-descriptor-1.4.5
django-celery-beat-2.7.0
django-timezone-field-7.0
python-crontab-3.2.0
